+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 - Rust 1.66.0 is now the earliest officially supported Rust version.
 - Update OpenSSL bindings and other crate versions to support latest Rust
   version.
+- Added option to redirect standard error to a file so that fatal errors do not
+  get sent over the wire to the client.